chore: 정해진 status code에 대해서는 StandardErrorPage로 분리
