+++ src/app/error/page.tsx
@@ -1,7 +1,7 @@
-import { ErrorPageLayout } from '@/features/customErrors/components';
+import { StandardErrorPage } from '@/features/customErrors/components';
 
 const ErrorPage = () => {
-  return <ErrorPageLayout statusCode={500} />;
+  return <StandardErrorPage statusCode={500} />;
 };
 
 export default ErrorPage;
