+++ src/app/home/[username]/error.tsx
@@ -1,9 +1,9 @@
 'use client';
 
-import { ErrorPageLayout } from '@/features/customErrors/components';
+import { StandardErrorPage } from '@/features/customErrors/components';
 
 const Error = ({ error }: { error: Error & { digest?: string } }) => {
-  return <ErrorPageLayout statusCode={403} />;
+  return <StandardErrorPage statusCode={403} />;
 };
 
 export default Error;
